+++ src/haddock/core/defaults.py
