+++ src/haddock/libs/libworkflow.py
@@ -5,7 +5,7 @@ import shutil
 from pathlib import Path
 from haddock.error import HaddockError, StepError
 # unused
-# from haddock.defaults import MODULE_PATH_NAME, TOPOLOGY_PATH
+# from haddock.core.defaults import MODULE_PATH_NAME, TOPOLOGY_PATH
 from haddock.modules import modules_category
 from haddock.libs.libutil import zero_fill
